Rec0vering - Unc0ver Removal Tool for iOS 11
ng is a unc0ver removal tool for iOS 11-11.4b3

- rec0vering's code will improve in the future. I an currently working with someone to improve on it..

- TWEAKS : tweak removal will be added soon. its in testing stages..

- NON JB DEVICES : The app version is in production right now. I'm not sure if you can use FilzaJailed or Escaped to achieve this
since I never tried this. 

COPIED SCRIPT : https://github.com/FivePixels/rec0vering- Has remade (with my help. We were working on remaking rec0vering in which
I helped with over half of the code, but he then said he was gonna call it his own) and called this his when it never was.
He isn't giving me credit (which is required in rec0verings license) and has A LOT of my code in this. He never tested if this worked
either so I suggest not using this for now..

Original rec0vering script by Chr1sPwn3d (@Chr1sPwn3d on twitter) and (Chr1sPwn3d#9333 on Discord)
NOT ASSOCIATED with unc0ver or the unc0ver team which includes pwn20wnd, Samg, SBingner and others.